add more asserts
git-svn-id: https://svn.circlemud.org/svn/rocket/trunk@640 46170a15-b891-4bf7-af97-fd374ce181e0

+++ src/lib/sim.c
@@ -382,7 +382,7 @@ void hal_uart_start_send()
 	LOGF((logfp, "Sent to uart: '%s'\n", buf));
 }
 
-void hal_uart_init(UartHandler *s, uint16_t baud, r_bool stop2)
+void hal_uart_init(UartHandler *s, uint32_t baud, r_bool stop2)
 {
 	g_sim_uart_handler = s;
 }
